+++ app/src/main/java/com/example/dentalhistoryrecorder/MainActivity.java
@@ -167,7 +167,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onResponse(String response) {
                 JSONArray jsonArray = null;
-                Toast.makeText(getApplicationContext(), response, Toast.LENGTH_LONG).show();
+//                Toast.makeText(getApplicationContext(), response, Toast.LENGTH_LONG).show();
             }
         }, new Response.ErrorListener() {
             @Override
